feat(unit-test): 전진 조건을 만족할 때 true/false가 나오는지 테스트
- 어떤 전략 패턴을 사용할 지 주입하는 구문 추가
